refactor(cli): format helpers
